+++ src/whisper-utils/whisper-processing.cpp
@@ -184,7 +184,7 @@ struct DetectionResultWithText run_whisper_inference(struct transcription_filter
 	if (gf->n_context_sentences > 0 && !gf->last_transcription_sentence.empty()) {
 		// set the initial prompt to the last transcription sentences (concatenated)
 		std::string initial_prompt = gf->last_transcription_sentence[0];
-		for (int i = 1; i < gf->last_transcription_sentence.size(); ++i) {
+		for (size_t i = 1; i < gf->last_transcription_sentence.size(); ++i) {
 			initial_prompt += " " + gf->last_transcription_sentence[i];
 		}
 		gf->whisper_params.initial_prompt = initial_prompt.c_str();
